Making the assertions in AboutAsserts.fs pass
What did I learn?...

'let' binds an identifier to a value. Note, that __ has already been bound to a value in Helpers.fs...
	let __ = "FILL ME IN"

It's statically typed though - the types are inferred (sometimes you need to add explicit type annotations if the compiler doesn't have enough information to infer it...I'm assuming there'll be more about this later)

You call a function by specifying the function name followed by a space and then list any arguments separated by spaces.
	AssertEquality (1 + 1) 2

(Also, the order of the files in the project matter...PathToElightenment becomes the entry point by way of being the last file. There's an attribute to explicitly marks an entry point, but I'm going to ignore that for now.)